// apps/inatrace-fe/src/api/model/apiResponseApiStockOrder.ts
export interface ApiResponseApiStockOrder {
  /**
   * Response status. OK for successful reponses.
   */
  status: ApiResponseApiStockOrder.StatusEnum;
  /**
   * Simple message to explain client developers the reason for error.
   */
  errorMessage?: string;
  data?: ApiStockOrder;
  /**
   * Optional details for unexpected error responses.
   */
  errorDetails?: string;
  validationErrorDetails?: ApiValidationErrorDetails;
}

/**
 * Namespace for property- and property-value-enumerations of ApiResponseApiStockOrder.
 */
export namespace ApiResponseApiStockOrder {
  /**
   * All properties of ApiResponseApiStockOrder.
   */
  export enum Properties {
    /**
     * Response status. OK for successful reponses.
     */
    status = 'status',
    /**
     * Simple message to explain client developers the reason for error.
     */
    errorMessage = 'errorMessage',
    data = 'data',
    /**
     * Optional details for unexpected error responses.
     */
    errorDetails = 'errorDetails',
    validationErrorDetails = 'validationErrorDetails',
  }

  /**
   * All possible values of status.
   */
  export enum StatusEnum {
    OK = 'OK',
    ERROR = 'ERROR',
    REQUESTBODYERROR = 'REQUEST_BODY_ERROR',
    VALIDATIONERROR = 'VALIDATION_ERROR',
    TOOMANYREQUESTS = 'TOO_MANY_REQUESTS',
    UNAUTHORIZED = 'UNAUTHORIZED',
    AUTHERROR = 'AUTH_ERROR',
    UPSTREAMHTTPERROR = 'UPSTREAM_HTTP_ERROR',
    INVALIDREQUEST = 'INVALID_REQUEST',
    INVALIDOREXPIREDSTORAGEKEY = 'INVALID_OR_EXPIRED_STORAGE_KEY',
    NOTIMPLEMENTED = 'NOT_IMPLEMENTED',
    NOTFOUND = 'NOT_FOUND',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiResponseApiStockOrder',
      vars: [
        {
          isReadOnly: false,
          isEnum: true,
          datatypeWithEnum: 'ApiResponseApiStockOrder.StatusEnum',
          required: true,
          name: 'status',
          classname: 'ApiResponseApiStockOrder',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'errorMessage',
          classname: 'ApiResponseApiStockOrder',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiStockOrder.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'data',
          classname: 'ApiResponseApiStockOrder',
          dataType: 'ApiStockOrder',
          isPrimitiveType: false,
          isListContainer: false,
          complexType: 'ApiStockOrder',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'errorDetails',
          classname: 'ApiResponseApiStockOrder',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiValidationErrorDetails.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'validationErrorDetails',
          classname: 'ApiResponseApiStockOrder',
          dataType: 'ApiValidationErrorDetails',
          isPrimitiveType: false,
          isListContainer: false,
          complexType: 'ApiValidationErrorDetails',
        },
      ],
      validators: {
        status: [['required']],
        errorMessage: [],
        data: [],
        errorDetails: [],
        validationErrorDetails: [],
      },
    };
  }

  // export const ApiResponseApiStockOrderValidationScheme = {
  //     validators: [],
  //     fields: {
  //               status: {
  //                   validators: []
  //               },
  //               errorMessage: {
  //                   validators: []
  //               },
  //               data: {
  //                   validators: []
  //               },
  //               errorDetails: {
  //                   validators: []
  //               },
  //               validationErrorDetails: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiResponseApiStockOrder>;
}